Save work mosty.
Changing noise generation to try and use a shaping function to force the edges to be water.
But that's buggy.
Add the voronoi_heightmap generator based on Azgaar's blog posts.
Trying to refactor main.rs a little to add options on the map generation type.
I don't think I understand how to use tilemaps.